Revert "[WOOMOB-2816] Add small home-screen layout to Store Stats widget"
